separated deconvolution layer file
